ndctl.spec.in: enable libtrace{event|fs} support for Fedora
As noted in https://src.fedoraproject.org/rpms/ndctl/pull-request/2,
the expression "0%{?rhel}" evaluates to zero on Fedora, so the
conditional "%if 0%{?rhel} < 9" evaluates to true, since 0 is less
than 9. The result is that ndctl builds for Fedora lack support for
libtraceevent and libtracefs. Correct the expression.

Reposted here from github pull request:
https://github.com/pmem/ndctl/pull/266/

Signed-off-by: Jerry James <loganjerry@gmail.com>
Reviewed-by: Vishal Verma <vishal.l.verma@intel.com>
Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/20240821214529.96966-1-alison.schofield@intel.com/
Signed-off-by: Alison Schofield <alison.schofield@intel.com>