test/daxctl-create.sh: use bash math syntax to find available size
The check for 1GB of available space in a DAX region always returned
true due to being wrapped inside a [[ ... ]] test, even when space
wasn't available. That caused set size to fail.

Update to use bash arithmetic evaluation instead.

This issue likely went unnoticed because users allocated >= 1GB of
efi_fake_mem. This fix is part of the transition to use CXL regions
in this test as efi_fake_mem support is being removed from the kernel.

Reviewed-by: Vishal Verma <vishal.l.verma@intel.com>a
Link: https://lore.kernel.org/r/865e28870eb8c072c2e368362a6d86fc4fb9cb61.1724813664.git.alison.schofield@intel.com/
Signed-off-by: Alison Schofield <alison.schofield@intel.com>