util/strbuf: remove unused cli infrastructure imports
The ndctl cli interface is built around an imported perf cli
infrastructure which was originally from git. [1]

A recent static analysis scan exposed an integer overflow issue in
strbuf_read() and although that is fixable, the function is not used
in ndctl. Further examination revealed additional unused functionality
in the string buffer handling import and a subset of that has already
been obsoleted from the perf cli.

In the interest of not maintaining unused code, remove the unused code
in util/strbuf.h,c. Ndctl, including cxl-cli and daxctl, are mature
cli's so it seems ok to let this functionality go after ~10 years.

In the interest of not touching what is not causing an issue, the
entirety of the original import was not reviewed at this time.

[1] 91677390f9e6 ("ndctl: import cli infrastructure from perf")

Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Link: https://lore.kernel.org/r/20250212034020.1865719-1-alison.schofield@intel.com
Signed-off-by: Alison Schofield <alison.schofield@intel.com>